datasetConstructor/requirements.txt
chardet>=5.2.0
markdown-it-py>=3.0.0
pydantic<2,>=1.10
python-slugify>=8.0.4
langid>=1.1.6
pandas>=2.2.0
pyarrow>=16.0.0
pyyaml>=6.0.0
pandas
pyarrow
markdown-it-py
langid
chardet
python-slugify
